services/model-registry/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9  # For file uploads

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Machine Learning
torch==2.2.0
onnx==1.15.0
onnxruntime==1.17.0

# Storage
google-cloud-storage==2.14.0  # GCS for model files
boto3==1.34.51
minio==7.2.4

# Monitoring
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1

# gRPC
grpcio==1.60.1
grpcio-tools==1.60.1
protobuf==4.25.2

# Development
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
black==24.2.0
ruff==0.2.2
mypy==1.8.0
